requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
dj-rest-auth==6.0.0
Django==5.1.2
django-admin-interface==0.29.1
django-allauth==65.1.0
django-ckeditor==6.7.1
django-cleanup==9.0.0
django-colorfield==0.11.0
django-cors-headers==4.6.0
django-js-asset==2.2.0
django-modeltranslation==0.19.11
django-rosetta==0.10.1
djangorestframework==3.15.2
idna==3.10
joblib==1.4.2
nltk==3.9.1
numpy==2.1.3
pandas==2.2.3
pillow==11.0.0
polib==1.2.0
psycopg2==2.9.10
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
pytz==2024.2
regex==2024.11.6
requests==2.32.3
rosetta==0.3
scipy==1.14.1
six==1.16.0
sqlparse==0.5.1
text-unidecode==1.3
tqdm==4.67.0
tzdata==2024.2
urllib3==2.2.3
